+++ src/components/molecules/RichMenulist/type.ts
@@ -1,0 +1,2 @@
+export type RichMenulistProps = {
+}
